Combininb delegate data source and tree renderer
