+++ src/vs/workbench/contrib/chat/browser/chatSessions/view/sessionsTreeRenderer.ts
@@ -8,7 +8,7 @@ import { $, append, getActiveWindow } from '../../../../../../base/browser/dom.j
 import { StandardKeyboardEvent } from '../../../../../../base/browser/keyboardEvent.js';
 import { ActionBar } from '../../../../../../base/browser/ui/actionbar/actionbar.js';
 import { InputBox, MessageType } from '../../../../../../base/browser/ui/inputbox/inputBox.js';
-import { ITreeNode, ITreeRenderer } from '../../../../../../base/browser/ui/tree/tree.js';
+import { IAsyncDataSource, ITreeNode, ITreeRenderer } from '../../../../../../base/browser/ui/tree/tree.js';
 import { timeout } from '../../../../../../base/common/async.js';
 import { Codicon } from '../../../../../../base/common/codicons.js';
 import { FuzzyScore } from '../../../../../../base/common/filters.js';
@@ -37,14 +37,16 @@ import { IResourceLabel, ResourceLabels } from '../../../../../browser/labels.js
 import { IEditableData } from '../../../../../common/views.js';
 import { IEditorGroupsService } from '../../../../../services/editor/common/editorGroupsService.js';
 import { IChatService } from '../../../common/chatService.js';
-import { ChatSessionStatus, IChatSessionItem, IChatSessionsService } from '../../../common/chatSessionsService.js';
+import { ChatSessionStatus, IChatSessionItem, IChatSessionItemProvider, IChatSessionsService } from '../../../common/chatSessionsService.js';
 import { ChatConfiguration } from '../../../common/constants.js';
 import { IChatWidgetService } from '../../chat.js';
 import { allowedChatMarkdownHtmlTags } from '../../chatMarkdownRenderer.js';
-import { ChatSessionItemWithProvider, getSessionItemContextOverlay, isLocalChatSessionItem } from '../../chatSessions/common.js';
+import { ChatSessionItemWithProvider, extractTimestamp, getSessionItemContextOverlay, isLocalChatSessionItem, processSessionsWithTimeGrouping } from '../common.js';
 import '../../media/chatSessions.css';
-import { LocalChatSessionsProvider } from '../../chatSessions/localChatSessionsProvider.js';
+import { LocalChatSessionsProvider } from '../localChatSessionsProvider.js';
 import { IListRenderer, IListVirtualDelegate } from '../../../../../../base/browser/ui/list/list.js';
+import { ChatSessionTracker } from '../chatSessionTracker.js';
+import { CancellationToken } from '../../../../../../base/common/cancellation.js';
 
 interface ISessionTemplateData {
 	readonly container: HTMLElement;
@@ -520,5 +522,118 @@ export class SessionsRenderer extends Disposable implements ITreeRenderer<IChatS
 		templateData.elementDisposable.dispose();
 		templateData.resourceLabel.dispose();
 		templateData.actionBar.dispose();
+	}
+}
+
+// Chat sessions item data source for the tree
+export class SessionsDataSource implements IAsyncDataSource<IChatSessionItemProvider, ChatSessionItemWithProvider> {
+
+	constructor(
+		private readonly provider: IChatSessionItemProvider,
+		private readonly chatService: IChatService,
+		private readonly sessionTracker: ChatSessionTracker,
+	) {
+	}
+
+	hasChildren(element: IChatSessionItemProvider | ChatSessionItemWithProvider): boolean {
+		const isProvider = element === this.provider;
+		if (isProvider) {
+			// Root provider always has children
+			return true;
+		}
+
+		// Check if this is the "Show history..." node
+		if ('id' in element && element.id === LocalChatSessionsProvider.HISTORY_NODE_ID) {
+			return true;
+		}
+
+		return false;
+	}
+
+	async getChildren(element: IChatSessionItemProvider | ChatSessionItemWithProvider): Promise<ChatSessionItemWithProvider[]> {
+		if (element === this.provider) {
+			try {
+				const items = await this.provider.provideChatSessionItems(CancellationToken.None);
+				const itemsWithProvider = items.map(item => {
+					const itemWithProvider: ChatSessionItemWithProvider = { ...item, provider: this.provider };
+
+					// Extract timestamp using the helper function
+					itemWithProvider.timing = { startTime: extractTimestamp(item) ?? 0 };
+
+					return itemWithProvider;
+				});
+
+				// Add hybrid local editor sessions for this provider using the centralized service
+				if (this.provider.chatSessionType !== 'local') {
+					const hybridSessions = await this.sessionTracker.getHybridSessionsForProvider(this.provider);
+					itemsWithProvider.push(...(hybridSessions as ChatSessionItemWithProvider[]));
+				}
+
+				// For non-local providers, apply time-based sorting and grouping
+				if (this.provider.chatSessionType !== 'local') {
+					processSessionsWithTimeGrouping(itemsWithProvider);
+				}
+
+				return itemsWithProvider;
+			} catch (error) {
+				return [];
+			}
+		}
+
+		// Check if this is the "Show history..." node
+		if ('id' in element && element.id === LocalChatSessionsProvider.HISTORY_NODE_ID) {
+			return this.getHistoryItems();
+		}
+
+		// Individual session items don't have children
+		return [];
+	}
+
+	private async getHistoryItems(): Promise<ChatSessionItemWithProvider[]> {
+		try {
+			// Get all chat history
+			const allHistory = await this.chatService.getHistory();
+
+			// Create history items with provider reference and timestamps
+			const historyItems = allHistory.map((historyDetail: any): ChatSessionItemWithProvider => ({
+				id: historyDetail.sessionId,
+				label: historyDetail.title,
+				iconPath: Codicon.chatSparkle,
+				provider: this.provider,
+				timing: {
+					startTime: historyDetail.lastMessageDate ?? Date.now()
+				},
+				isHistory: true,
+			}));
+
+			// Apply sorting and time grouping
+			processSessionsWithTimeGrouping(historyItems);
+
+			return historyItems;
+
+		} catch (error) {
+			return [];
+		}
+	}
+}
+
+
+export class SessionsDelegate implements IListVirtualDelegate<ChatSessionItemWithProvider> {
+	static readonly ITEM_HEIGHT = 22;
+	static readonly ITEM_HEIGHT_WITH_DESCRIPTION = 44; // Slightly smaller for cleaner look
+
+	constructor(private readonly configurationService: IConfigurationService) { }
+
+	getHeight(element: ChatSessionItemWithProvider): number {
+		// Return consistent height for all items (single-line layout)
+		if (element.description && this.configurationService.getValue(ChatConfiguration.ShowAgentSessionsViewDescription) && element.provider.chatSessionType !== 'local') {
+			return SessionsDelegate.ITEM_HEIGHT_WITH_DESCRIPTION;
+		} else {
+			return SessionsDelegate.ITEM_HEIGHT;
+		}
+	}
+
+	getTemplateId(element: ChatSessionItemWithProvider): string {
+		return SessionsRenderer.TEMPLATE_ID;
 	}
 }
